requirements.txt
streamlit
openai
azure-cognitiveservices-speech
pandas
python-dotenv
PyMuPDF
streamlit-webrtc
requests
